chore(test): Add configurable opts to e2e tests (#239)
* Add configurable opts to e2e tests
* Only run e2e nightly if on this repo and pr
* Bump timeout for dmsetup

I have added the initial frame for custom test flags.
So far just the options to:
- Skip setting up the devmapper thinpool
- Skip deleting VMs
- Skip all teardown steps
- Set log level for containerd
- Set log level for flintlockd

Unfortunately custom test flags can only be added on `init()` which is
bleh but I don't see another nice was of doing this.

These flags can only be used while running `go test ...` directly.
This means you cannot pass in flags as part of a `make` or a
`docker run -it ...` command, but you can just call `go test etc` from inside the
container or in the metal host and configure what you like.

My next step is some more work around the python tooling so that these
(and more) options can be passed up.

+++ test/e2e/e2e_test.go
@@ -13,6 +13,15 @@ import (
 	u "github.com/weaveworks/flintlock/test/e2e/utils"
 )
 
+var params *u.Params
+
+func init() {
+	// Call testing.Init() prior to tests.NewParams(), as otherwise custom test flags
+	// will not be recognised.
+	testing.Init()
+	params = u.NewParams()
+}
+
 func TestE2E(t *testing.T) {
 	RegisterTestingT(t)
 
@@ -26,7 +35,7 @@ func TestE2E(t *testing.T) {
 		mvmPid2 int
 	)
 
-	r := u.Runner{}
+	r := u.NewRunner(params)
 	defer func() {
 		log.Println("TEST STEP: cleaning up running processes")
 		r.Teardown()
@@ -78,6 +87,11 @@ func TestE2E(t *testing.T) {
 		g.Expect(res.Microvm[1].Status.State).To(Equal(types.MicroVMStatus_CREATED))
 		return nil
 	}, "120s").Should(Succeed())
+
+	if params.SkipDelete {
+		log.Println("TEST STEP: skipping delete")
+		return
+	}
 
 	log.Println("TEST STEP: deleting existing MicroVMs")
 	Expect(u.DeleteMVM(flintlockClient, mvmID, mvmNS)).To(Succeed())
